requirements.txt
# ── TTS ────────────────────────────────────────────────────────────────
TTS>=0.22.0                  # Coqui XTTS-v2 (pip install TTS)

# ── Face enhancement ────────────────────────────────────────────────────
gfpgan>=1.3.8

# ── Video / image processing ─────────────────────────────────────────
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
numpy>=1.24.0

# ── Cloud / storage ──────────────────────────────────────────────────
boto3>=1.34.0
supabase>=2.4.0

# ── Utilities ────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0

# ── GPU ──────────────────────────────────────────────────────────────
# Install PyTorch separately with the correct CUDA build for your GPU:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0

# ── NOT pip-installable — clone separately ───────────────────────────
# MuseTalk  : git clone https://github.com/TMElyralab/MuseTalk deps/MuseTalk
# SadTalker : git clone https://github.com/OpenTalker/SadTalker deps/SadTalker
# Set MUSETALK_DIR / SADTALKER_DIR in .env after cloning
